Extract SearchService from search controller
