+++ convergence_games/apps/frontend/games/controllers/_search.py
@@ -1,16 +1,7 @@
-from dataclasses import dataclass
-
 from litestar import Controller, get
+from litestar.di import Provide
 from litestar.exceptions import NotFoundException
-from rapidfuzz import fuzz, process, utils
-from sqlalchemy import select
-from sqlalchemy.ext.asyncio import AsyncSession
-from sqlalchemy.orm import selectinload
-from sqlalchemy.sql import ColumnExpressionArgument
 
-from convergence_games.db.enums import (
-    SubmissionStatus,
-)
 from convergence_games.db.models import (
     ContentWarning,
     Genre,
@@ -20,89 +11,12 @@ from convergence_games.lib.ocean import Sqid, sink
 from convergence_games.lib.request_type import Request
 from convergence_games.lib.response_type import HTMXBlockTemplate, Template
 
-type SearchableBase = System | Genre | ContentWarning
-
-
-@dataclass
-class SearchResult[T: SearchableBase]:
-    name: str
-    match: str
-    score: float
-    result: T
-
-
-async def search_with_fuzzy_match[T: SearchableBase](
-    transaction: AsyncSession,
-    model_type: type[T],
-    search: str,
-    extra_filters: ColumnExpressionArgument[bool] | None = None,
-    suggested_on_empty: bool = False,
-) -> list[SearchResult[T]]:
-    # TODO: Can we directly query for the names (possibly including aliases) and sqids?
-    query = select(model_type)
-
-    if issubclass(model_type, System):
-        query = query.options(selectinload(model_type.aliases))
-
-    if extra_filters is not None:
-        query = query.where(extra_filters)
-
-    if not search:
-        if suggested_on_empty:
-            assert issubclass(model_type, Genre) or issubclass(model_type, ContentWarning)
-            query = query.where(model_type.suggested).order_by(model_type.name)
-            all_rows = (await transaction.execute(query)).scalars().all()
-            return [
-                SearchResult(
-                    name=row.name,
-                    match=row.name,
-                    score=100,
-                    result=row,
-                )
-                for row in all_rows
-            ]
-        else:
-            return []
-
-    all_rows = (await transaction.execute(query)).scalars().all()
-
-    to_match: list[tuple[str, T]] = []
-    for row in all_rows:
-        to_match.append((row.name, row))
-        if isinstance(row, System):
-            for alias in row.aliases:
-                to_match.append((alias.name, row))
-
-    names_scores_indices = process.extract(
-        query=search,
-        choices=[name for name, _ in to_match],
-        scorer=fuzz.WRatio,
-        processor=utils.default_process,
-        limit=10,
-        score_cutoff=50,
-    )
-
-    already_matched_ids: set[int] = set()
-    top_results: list[SearchResult[T]] = []
-
-    for _, score, index in names_scores_indices:
-        result = to_match[index][1]
-        if result.id not in already_matched_ids:
-            top_results.append(
-                SearchResult(
-                    name=result.name,
-                    match=to_match[index][0],
-                    score=score,
-                    result=result,
-                )
-            )
-            already_matched_ids.add(result.id)
-
-    return top_results
+from ..services import SearchService, provide_search_service
 
 
 class SearchController(Controller):
     path = "/search"
+    dependencies = {"search_service": Provide(provide_search_service)}
 
     @get(path="/{name:str}")
     async def get_search(self, name: str) -> Template:
@@ -118,11 +32,8 @@ class SearchController(Controller):
         )
 
     @get(path="/system/results")
-    async def get_system_search_results(self, request: Request, transaction: AsyncSession, search: str) -> Template:
-        extra_filters = System.submission_status == SubmissionStatus.APPROVED
-        if request.user:
-            extra_filters = extra_filters | (System.created_by == request.user.id)
-        results = await search_with_fuzzy_match(transaction, System, search, extra_filters)
+    async def get_system_search_results(self, request: Request, search_service: SearchService, search: str) -> Template:
+        results = await search_service.search_visible_to_user(System, search, request.user.id if request.user else None)
 
         return HTMXBlockTemplate(
             template_name="components/forms/search/SearchResultsList.html.jinja",
@@ -135,8 +46,8 @@ class SearchController(Controller):
         )
 
     @get(path="/system/select")
-    async def get_system_search_selected(self, transaction: AsyncSession, sqid: Sqid) -> Template:
-        system = await transaction.get(System, sink(sqid))
+    async def get_system_search_selected(self, search_service: SearchService, sqid: Sqid) -> Template:
+        system = await search_service.get_by_id(System, sink(sqid))
 
         if not system:
             raise NotFoundException(detail="System not found")
@@ -162,11 +73,10 @@ class SearchController(Controller):
         )
 
     @get(path="/genre/results")
-    async def get_genre_search_results(self, request: Request, transaction: AsyncSession, search: str) -> Template:
-        extra_filters = Genre.submission_status == SubmissionStatus.APPROVED
-        if request.user:
-            extra_filters = extra_filters | (Genre.created_by == request.user.id)
-        results = await search_with_fuzzy_match(transaction, Genre, search, extra_filters, suggested_on_empty=True)
+    async def get_genre_search_results(self, request: Request, search_service: SearchService, search: str) -> Template:
+        results = await search_service.search_visible_to_user(
+            Genre, search, request.user.id if request.user else None, suggested_on_empty=True
+        )
 
         return HTMXBlockTemplate(
             template_name="components/forms/search/SearchResultsList.html.jinja",
@@ -179,8 +89,8 @@ class SearchController(Controller):
         )
 
     @get(path="/genre/select")
-    async def get_genre_search_selected(self, transaction: AsyncSession, sqid: Sqid) -> Template:
-        genre = await transaction.get(Genre, sink(sqid))
+    async def get_genre_search_selected(self, search_service: SearchService, sqid: Sqid) -> Template:
+        genre = await search_service.get_by_id(Genre, sink(sqid))
 
         if not genre:
             raise NotFoundException(detail="Genre not found")
@@ -207,13 +117,10 @@ class SearchController(Controller):
 
     @get(path="/content-warning/results")
     async def get_content_warning_search_results(
-        self, request: Request, transaction: AsyncSession, search: str
+        self, request: Request, search_service: SearchService, search: str
     ) -> Template:
-        extra_filters = ContentWarning.submission_status == SubmissionStatus.APPROVED
-        if request.user:
-            extra_filters = extra_filters | (ContentWarning.created_by == request.user.id)
-        results = await search_with_fuzzy_match(
-            transaction, ContentWarning, search, extra_filters, suggested_on_empty=True
+        results = await search_service.search_visible_to_user(
+            ContentWarning, search, request.user.id if request.user else None, suggested_on_empty=True
         )
 
         return HTMXBlockTemplate(
@@ -227,8 +134,8 @@ class SearchController(Controller):
         )
 
     @get(path="/content-warning/select")
-    async def get_content_warning_search_selected(self, transaction: AsyncSession, sqid: Sqid) -> Template:
-        content_warning = await transaction.get(ContentWarning, sink(sqid))
+    async def get_content_warning_search_selected(self, search_service: SearchService, sqid: Sqid) -> Template:
+        content_warning = await search_service.get_by_id(ContentWarning, sink(sqid))
 
         if not content_warning:
             raise NotFoundException(detail="Content warning not found")
