+++ convergence_games/apps/frontend/games/services/__init__.py
@@ -1,0 +1,8 @@
+from ._search_service import SearchableBase, SearchResult, SearchService, provide_search_service
+
+__all__ = [
+    "SearchResult",
+    "SearchService",
+    "SearchableBase",
+    "provide_search_service",
+]
